% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addQC.R
\name{addQC}
\alias{addQC}
\title{Include QC samples in sample sequence}
\usage{
addQC(
  samples = NULL,
  interval = NULL,
  position = NULL,
  replace = FALSE,
  permute = FALSE,
  label = "QC"
)
}
\arguments{
\item{samples}{A vector containing sample names}

\item{interval}{An interval at which QC samples should be inserted. Specify only either \code{interval} or \code{position}.}

\item{position}{A vector of positions at which QC samples should be inserted. The position can contain duplicated value, it will add QC next to each other. Specify only either \code{interval} or \code{position}.}

\item{replace}{If \code{TRUE} and \code{position} is specified, samples at the given
positions are replaced by QC labels in-place (sequence length unchanged).
If \code{FALSE} (default), QC labels are inserted as additional entries
adjacent to the specified positions, increasing the sequence length.
\code{replace} is ignored when \code{interval} is used.}

\item{permute}{If TRUE will randomize an order within each QC block. Can only be used with \code{interval}.}

\item{label}{Prefix of added QC}
}
\value{
A vector containing added QC
}
\description{
Include QC samples by repeated interval or at specified positions in a sample sequence.
}
\examples{
samples <- LETTERS[1:20]
addQC(samples, interval = 5)
addQC(samples, interval = 5, permute = TRUE)
addQC(samples, position = c(3, 9), replace = TRUE)
addQC(samples, position = c(rep(3,5), rep(9,10)))
}
\seealso{
Other plate: 
\code{\link{fill.plate}()},
\code{\link{generate.position}()},
\code{\link{merge.plate_list}()},
\code{\link{plot.plate}()},
\code{\link{plot.plate_list}()},
\code{\link{print.plate_list}()},
\code{\link{write.sld}()}
}
\concept{plate}
