% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{ccc}
\alias{ccc}
\title{Example clinical covariates dataset}
\format{
A data frame with sample metadata columns including:
\describe{
\item{sample}{Sample identifier}
\item{time}{Time-point covariate}
\item{response}{Response-group covariate}
}
}
\usage{
data(ccc)
}
\description{
A small example data frame used to demonstrate block randomisation.
}
\keyword{datasets}
