% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.block.R
\name{check.block}
\alias{check.block}
\title{Check block distribution}
\usage{
check.block(df, block = block.id, group = unique_group)
}
\arguments{
\item{df}{An output object form block.rand()}

\item{block}{A column containing block id}

\item{group}{A column containing unique covariates}
}
\value{
A list of table contains distribution of covariates in each block.
}
\description{
Check a covariates distribution in each block.
}
\examples{
data(ccc)
ccc_block <- block.rand(ccc, time, response)
check.block(ccc_block)

}
\seealso{
Other randomize: 
\code{\link{block.rand}()}
}
\concept{randomize}
