% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIA.windows.R
\name{DIA.windows}
\alias{DIA.windows}
\title{Generate DIA windows}
\usage{
DIA.windows(
  mnc = NULL,
  mxc = NULL,
  windows.width = NULL,
  n.windows = NULL,
  optimize = TRUE,
  stagger = FALSE,
  margin = 0,
  output.type = "range",
  save.csv = FALSE
)
}
\arguments{
\item{mnc}{A minimal cap of GPF windows}

\item{mxc}{A maximal cap of GPF windows}

\item{windows.width}{Isolation width. Specify only either \code{windows.width} or \code{n.windows}.}

\item{n.windows}{Number of windows. Specify only either \code{windows.width} or \code{n.windows}.}

\item{optimize}{If \code{TRUE} will place the edge on the peptides forbidden zones.}

\item{stagger}{If \code{TRUE} will generated series of staggered windows.}

\item{margin}{An isolation margin}

\item{output.type}{choose from \code{"range"}, \code{"center"}, \code{"both"}}

\item{save.csv}{If \code{TRUE} also write output to \code{.csv} files}
}
\value{
A dataframe containing m/z sequential windows for data-independent acquisition
\itemize{
\item If \code{output.type = "range"} will show an output as min-max range.
\item If \code{output.type = "center"} will show an output as center mass and scan width.
\item If \code{output.type = "both"} will show all columns mentioned earlier.
}
}
\description{
This function generate sequential windows for data-independent acquisition.
}
\examples{
DIA.windows(mnc = 400, mxc = 1000, windows.width = 20)
DIA.windows(mnc = 400, mxc = 1000, windows.width = 20, output.type = "center")
DIA.windows(mnc = 400, mxc = 1000, windows.width = 20, stagger = TRUE)
DIA.windows(mnc = 400, mxc = 1000, windows.width = 20, optimize = TRUE, stagger = TRUE)
}
\seealso{
Other m/z: 
\code{\link{GPF.windows}()},
\code{\link{plot.mzdf}()},
\code{\link{print.mzdf}()},
\code{\link{vDIA.windows}()}
}
\concept{m/z}
