% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.inner.plate.R
\name{fill.inner.plate}
\alias{fill.inner.plate}
\title{Fill a single plate with data}
\usage{
fill.inner.plate(
  samples,
  nrows,
  ncols,
  row_letters,
  plate_names,
  exclude = NULL,
  exclude_label = "iQC"
)
}
\arguments{
\item{samples}{A vector of data to fill the plate}

\item{nrows}{Number of rows in a plate}

\item{ncols}{Number of columns in a plate}

\item{row_letters}{If \code{TRUE} will display \code{rownames} as alphabets}

\item{plate_names}{A vector of plate names to be used}

\item{exclude}{Well positions to reserve (character vector or named list)}

\item{exclude_label}{Label placed in reserved wells}
}
\value{
A list of matrices representing filled plates
}
\description{
This function fills a single plate with the provided data.
}
