% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.position.R
\name{generate.position}
\alias{generate.position}
\title{Generate plate position}
\usage{
generate.position(plate, flatten = FALSE)
}
\arguments{
\item{plate}{A list of plate matrices generated from fill.plate().}

\item{flatten}{If \code{TRUE} will flatten a list into a single vector.}
}
\value{
A vector or list of vectors containing \code{plate_name:plate_position}
}
\description{
This function retrieve plate matrices from \code{fill.plate()} and flatten into vector in the form that can be filled into MS sequence file.
}
\examples{
x_plate <- fill.plate(1:100)
plate_list <- generate.position(x_plate)
plate_list

## Use unlist to flatten the list into vectors
unlist(plate_list, recursive = FALSE, use.names = FALSE)

## Or use flatten = TRUE to flatten all results
all.equal(generate.position(x_plate, flatten = TRUE),
unlist(plate_list, recursive = TRUE, use.names = FALSE)) ## TRUE
}
\seealso{
Other plate: 
\code{\link{addQC}()},
\code{\link{fill.plate}()},
\code{\link{merge.plate_list}()},
\code{\link{plot.plate}()},
\code{\link{plot.plate_list}()},
\code{\link{print.plate_list}()},
\code{\link{write.sld}()}
}
\concept{plate}
