% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPF.windows.R
\name{GPF.windows}
\alias{GPF.windows}
\title{Make an isolation list for GPF-DIA}
\usage{
GPF.windows(
  isolation_list = NULL,
  lower_windows = NULL,
  upper_windows = NULL,
  windows_width = 100,
  mnc = NULL,
  mxc = NULL,
  output.type = "range",
  save.csv = FALSE
)
}
\arguments{
\item{isolation_list}{A vector of isolation list with range, e.g. \code{400-420}}

\item{lower_windows}{Lower windows in isolation list. Ignore if \code{isolation_list} is used.}

\item{upper_windows}{Upper windows in isolation list. Ignore if \code{isolation_list} is used.}

\item{windows_width}{GPF-DIA scan range per group, default = \code{100}.}

\item{mnc}{A minimal cap of GPF windows. If \code{NULL} will determine from the lowest value of lower bound.}

\item{mxc}{A maximal cap of GPF windows. If \code{NULL} will determine from the highest value of lower bound.}

\item{output.type}{choose from \code{"range"}, \code{"center"}, \code{"both"}}

\item{save.csv}{If \code{TRUE} also write output to \code{.csv} files}
}
\value{
A list of dataframe containing m/z separation windows for each GPF group.
\itemize{
\item If \code{output.type = "range"} will show an output as min-max range.
\item If \code{output.type = "center"} will show an output as center mass and scan width.
\item If \code{output.type = "both"} will show all columns mentioned earlier.
}
}
\description{
The function utilizes a predefined list of isolation windows and partitions them into groups for gas-phase fractionation DIA experiments.
}
\examples{
data(staggered_mz)
GPF.windows(staggered_mz$range)

## Can also use vector of lower and upper m/z range.
mz_list <- strsplit(staggered_mz$range, split = "-")
lower_bound <- as.numeric(sapply(mz_list, "[[", 1))
upper_bound <- as.numeric(sapply(mz_list, "[[", 2))
GPF.windows(lower_windows = lower_bound, upper_windows = upper_bound)
}
\seealso{
Other m/z: 
\code{\link{DIA.windows}()},
\code{\link{plot.mzdf}()},
\code{\link{print.mzdf}()},
\code{\link{vDIA.windows}()}
}
\concept{m/z}
