% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.windows.R
\name{opt.windows}
\alias{opt.windows}
\title{Optimized windows placement}
\usage{
opt.windows(mass, phospho = FALSE)
}
\arguments{
\item{mass}{A numeric vector containing edges of isolation windows}

\item{phospho}{If \code{TRUE} will also adjust for phosphopeptides.}
}
\value{
Adjusted masses
}
\description{
This function adjusts the edges of isolation windows to position them within forbidden zones, where +2H and +3H peptides are unlikely to be presented.
}
\examples{
windows <- seq(400,1000,20)
opt.windows(windows)
}
\references{
\itemize{
\item Egertson JD, Kuehn A, Merrihew GE, Bateman NW, MacLean BX, Ting YS, et al. Multiplexed MS/MS for improved data-independent acquisition. Nat Methods 2013;10:744–6.
\item Pino LK, Just SC, MacCoss MJ, Searle BC. Acquiring and Analyzing Data Independent Acquisition Proteomics Experiments without Spectrum Libraries. Molecular & Cellular Proteomics 2020;19:1088–103.
}
}
