% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.windows.R
\name{plot.mzdf}
\alias{plot.mzdf}
\title{Plot windows range.}
\usage{
\method{plot}{mzdf}(x, ...)
}
\arguments{
\item{x}{A dataframe with class \code{mzdf}}

\item{...}{Currently unused; included for S3 method compatibility.}
}
\value{
A \code{ggplot} object of windows range, can be further modified or saved by \code{ggplot2} function.
}
\description{
Plot windows range.
}
\examples{
dia_range <- DIA.windows(400,1000,20)
plot(dia_range)
}
\seealso{
Other m/z: 
\code{\link{DIA.windows}()},
\code{\link{GPF.windows}()},
\code{\link{print.mzdf}()},
\code{\link{vDIA.windows}()}
}
\concept{m/z}
