% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.plate.R
\name{plot.plate}
\alias{plot.plate}
\title{Plot a plate}
\usage{
\method{plot}{plate}(
  x,
  labels = NULL,
  text = TRUE,
  control.suffix = NULL,
  return_df = FALSE,
  ...
)
}
\arguments{
\item{x}{A plate matrix generated from fill.plate()}

\item{labels}{A vector of label in row order, can be ignored}

\item{text}{If \code{TRUE} will also display texts in the plate.}

\item{control.suffix}{Vectors of control, will remove their suffixes, which make them belong to the same group.}

\item{return_df}{If \code{TRUE} will also return a dataframe that is used to plot.}

\item{...}{Currently unused; included for S3 method compatibility.}
}
\value{
A \code{ggplot} object of a plate, can be further modified or saved by \code{ggplot2} function.
}
\description{
This function plot a plate colored by their groups.
}
\examples{
data(ccc)
block_df <- block.rand(ccc, time, response)
plate <- fill.plate(block_df$unique_group_id)
plot(plate$plate[[1]])
plot(plate$plate[[1]], labels = generate.position(plate)[[1]])
}
\seealso{
Other plate: 
\code{\link{addQC}()},
\code{\link{fill.plate}()},
\code{\link{generate.position}()},
\code{\link{merge.plate_list}()},
\code{\link{plot.plate_list}()},
\code{\link{print.plate_list}()},
\code{\link{write.sld}()}
}
\concept{plate}
