% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precursor_hela.R
\docType{data}
\name{precursor_hela}
\alias{precursor_hela}
\title{Hela precursor m/z density}
\format{
A data frame with precursor m/z and their frequency
\describe{
\item{mz}{Precursor m/z}
\item{n}{Detection frequency aggregated from all experiments}
}
}
\source{
Webel H, Perez-Riverol Y, Nielsen AB, Rasmussen S. Mass spectrometry-based proteomics data from thousands of HeLa control samples. Sci Data 2024;11:112.
}
\usage{
data(precursor_hela)
}
\description{
The data contains precursor frequency of hela protein digested summarized from Webel et al.
}
