% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.methods.R
\name{print.block.rand}
\alias{print.block.rand}
\title{Print method for block.rand objects}
\usage{
\method{print}{block.rand}(x, ...)
}
\arguments{
\item{x}{A \code{block.rand} object returned by \code{\link[=block.rand]{block.rand()}}.}

\item{...}{Further arguments passed to \code{\link[=NextMethod]{NextMethod()}}.}
}
\value{
\code{x} invisibly.
}
\description{
Shows a brief header then delegates to the data-frame print.
}
\concept{block}
