% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.methods.R
\name{print.mzdf}
\alias{print.mzdf}
\title{Print method for mzdf objects}
\usage{
\method{print}{mzdf}(x, ...)
}
\arguments{
\item{x}{A \code{mzdf} object returned by \code{\link[=DIA.windows]{DIA.windows()}}.}

\item{...}{Further arguments passed to \code{\link[=NextMethod]{NextMethod()}}.}
}
\value{
\code{x} invisibly.
}
\description{
Shows a brief header then delegates to the data-frame print.
}
\seealso{
Other m/z: 
\code{\link{DIA.windows}()},
\code{\link{GPF.windows}()},
\code{\link{plot.mzdf}()},
\code{\link{vDIA.windows}()}
}
\concept{m/z}
