% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.ext.R
\name{seq.ext}
\alias{seq.ext}
\title{Extend sequence to max value}
\usage{
\method{seq}{ext}(
  from = 1,
  to = 1,
  by = ((to - from)/(length.out - 1)),
  ...,
  rm.last = FALSE
)
}
\arguments{
\item{from}{\code{from} for seq}

\item{to}{\code{to} for seq}

\item{by}{\code{by} for seq}

\item{...}{other arguments passed to native \code{seq}}

\item{rm.last}{whether to remove last value before adding max value}
}
\value{
A sequence with extension to max value
}
\description{
This function extends a sequence to the maximum value and allows the option to remove the last element from the sequence.
}
\examples{
seq.ext(1,9,3)
seq.ext(1,9,3, rm.last = TRUE)
}
\concept{utils}
