% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vDIA.windows.R
\name{vDIA.windows}
\alias{vDIA.windows}
\title{Generate variable DIA windows}
\usage{
vDIA.windows(
  mnc = NULL,
  mxc = NULL,
  n.windows = NULL,
  optimize = TRUE,
  margin = 0,
  ref.data = NULL,
  output.type = "range",
  save.csv = FALSE
)
}
\arguments{
\item{mnc}{A minimal cap of GPF windows}

\item{mxc}{A maximal cap of GPF windows}

\item{n.windows}{Number of windows.}

\item{optimize}{If \code{TRUE} will place the edge on the peptides forbidden zones.}

\item{margin}{An isolation margin}

\item{ref.data}{Reference dataframe, should has integer \code{mz} and frequency \code{n} columns. If \code{NULL} will use \code{precursor_hela} data.}

\item{output.type}{choose from \code{"range"}, \code{"center"}, \code{"both"}}

\item{save.csv}{If \code{TRUE} also write output to \code{.csv} files}
}
\value{
A dataframe containing m/z sequential windows for data-independent acquisition
\itemize{
\item If \code{output.type = "range"} will show an output as min-max range.
\item If \code{output.type = "center"} will show an output as center mass and scan width.
\item If \code{output.type = "both"} will show all columns mentioned earlier.
}
}
\description{
This function generate variable sequential windows for data-independent acquisition. The windows are based on reference data provided.
}
\examples{
vDIA.windows(mnc = 400, mxc = 1000, n.windows = 32)
}
\seealso{
Other m/z: 
\code{\link{DIA.windows}()},
\code{\link{GPF.windows}()},
\code{\link{plot.mzdf}()},
\code{\link{print.mzdf}()}
}
\concept{m/z}
