% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.sld.R
\name{write.sld}
\alias{write.sld}
\title{Write a Thermo sequence file}
\usage{
write.sld(
  mat = NULL,
  filename = NULL,
  return_df = TRUE,
  samp_type = NULL,
  sample_name = NULL,
  path = "",
  inst_method = "",
  proc_meth = "",
  position = NULL,
  inj_vol = NULL
)
}
\arguments{
\item{mat}{A list of matrix from \code{fill.plate()}}

\item{filename}{File name, should be .csv}

\item{return_df}{If \code{TRUE} will also return the resulting dataframe to console.}

\item{samp_type}{If \code{NULL} will use data from \code{mat}. Automatically check for QC or blank in names.}

\item{sample_name}{If \code{NULL} will use data from plate content.}

\item{path}{Path to save files when run experiment in Xcalibur. Should have \verb{length = 1 or length(sample_name)}, can be ignored.}

\item{inst_method}{Path to instrument method. Should have \verb{length = 1 or length(sample_name)}, can be ignored.}

\item{proc_meth}{Path to processing method. Should have \verb{length = 1 or length(sample_name)}, can be ignored.}

\item{position}{If \code{NULL} will take plates' positions from \code{mat}}

\item{inj_vol}{Injection volumn in ul. Should have \verb{length = 1 or length(sample_name)}, can be ignored.}
}
\value{
A \code{.csv} file that can be imported to Xcalibur. Also return a dataframe if \code{return_df = TRUE}
}
\description{
This function take a plate list from \code{fill.plate()} and make a \code{.csv} file that can be imported to Thermo's Xcalibur as \code{.sld} sequence.
}
\examples{
write.sld(fill.plate(addQC(1:100,5)), filename = "test.csv")
}
\seealso{
Other plate: 
\code{\link{addQC}()},
\code{\link{fill.plate}()},
\code{\link{generate.position}()},
\code{\link{merge.plate_list}()},
\code{\link{plot.plate}()},
\code{\link{plot.plate_list}()},
\code{\link{print.plate_list}()}
}
\concept{plate}
